% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ai.R, R/get_equis_monitors.R
\name{get_ai}
\alias{get_ai}
\alias{get_ai_names}
\title{Get all Agency Interest (AI) names from TEMPO.}
\usage{
get_ai(ai = NULL, keep_alt_names = TRUE, tempo_test = FALSE)

get_ai_names(ai = NULL, keep_alt_names = TRUE, tempo_test = FALSE)

get_ai_names(ai = NULL, keep_alt_names = TRUE, tempo_test = FALSE)
}
\arguments{
\item{ai}{Vector of Agency Interest numbers to keep. The default, `NULL`. will keep all values.}

\item{keep_alt_names}{Include alternative names? Default is `TRUE`. which will include a facility's previous names.}

\item{tempo_test}{Connect to TEMPO's test database? Default is `FALSE`. which connects to TEMPO's working / production database.}
}
\description{
Requires a connection to MPCA's internal network.
}
\examples{
library(mpcadb)

/dontrun{
get_ai(ai = c(3421, 566), keep_alt_names = TRUE, tempo_test = FALSE)
}

get_ai_names(ai = c(4558, 12, 549, 1:5), keep_alt_names = TRUE, tempo_test = FALSE)
get_ai_names(ai = c(4558, 12, 549, 1:5), keep_alt_names = TRUE, tempo_test = FALSE)
}
