% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_creds.R
\name{get_creds}
\alias{get_creds}
\title{Get credentials for MPCA's internal oracle database DELTA}
\usage{
get_creds()
}
\description{
Requires a direct connection to MPCA's internal network. If connection exists, the function will generate login credentials.
}
\examples{
library(mpcadb)

/dontrun{
get_creds()
}

}
