% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_delta.R
\name{open_delta}
\alias{open_delta}
\title{Open a connection to MPCA's internal oracle database DELTAW, which houses TEMPO, EQUIS, RAPIDS, ONBASE}
\usage{
open_delta(delta_dsn = NULL)
}
\description{
Requires access to MPCA's internal network and X-drive.
}
\details{
#' @param delta_dsn The Data Source Name assigned to the `DELTAW` database. The default value, `NULL`. will search your computer's data sources for "delta" and use the first returned value.
}
\examples{
library(mpcadb)

/dontrun{
open_delta()
}

}
